# backend/internal/docs/swagger.yaml
basePath: /api/
definitions:
  _ResponseWithData:
    description: Общий ответ success/error, содержащий произвольные данные.
    properties:
      data:
        description: Объект полезной нагрузки
      status:
        description: Результат запроса
        type: string
    type: object
  _ResponseWithMessage:
    description: Общий простой ответ, который передает только понятное для человека
      сообщение.
    properties:
      message:
        description: Человеко-читаемое сообщение
        type: string
      status:
        description: Результат запроса
        type: string
    type: object
...
  ForgotPasswordRequest:
    description: Запрос на восстановление пароля.
    properties:
      email:
        description: Почта, на которую придёт письмо для восстановления пароля
        type: string
    required:
    - email
    type: object
...
  ResetPasswordRequest:
    description: Запрос на сброс пароля.
    properties:
      newPassword:
        description: Новый пароль
        type: string
      token:
        description: Токен, полученный по ссылке из письма
        type: string
    required:
    - newPassword
    - token
    type: object
...
  User:
    description: Модель пользователя, хз что ещё сказать можно по этому поводу.
    properties:
      blocked:
        description: Заблокирован ли пользователь
        example: false
        type: boolean
      confirmed:
        description: Подтверждён ли пользователь
        example: true
        type: boolean
      createdAt:
        description: Timestamp создания аккаунта
        example: "2006-01-02T15:04:05Z"
        format: date-time
        type: string
      deleted:
        description: Удалён ли пользователь
        example: true
        type: boolean
      email:
        description: Электронная почта пользователя
        example: Dimka228@gmail.com
        type: string
      id:
        description: ID пользователя
        example: b4b03119-1290-44bc-b599-6a5e91d6611f
        type: string
      role:
        description: Роль пользователя
        example: user
        type: string
      updatedAt:
        description: Timestamp последнего обновления аккаунта
        example: "2006-01-02T15:04:05Z"
        format: date-time
        type: string
      username:
        description: Имя пользователя
        example: Dimka228
        type: string
    required:
    - blocked
    - confirmed
    - createdAt
    - deleted
    - email
    - id
    - role
    - updatedAt
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Флоу авторизации: сначала пользователь регистрируется, в ответе получает модельку своего user и токен.
    Токен + код с почты он отправляет на ручку confirm, если что-то идёт не так (токен истёк, не пришёл код на почту), то запрос нужно отправить на ручку resend-confirmation.
    Далее уже можно авторизоваться, login/refresh/test-login выставляет в cookie access и refresh токены, фронтенду, ничего с ними делать не нужно, они сами по себе живут в браузере и отправляются при каждом запросе.
    Специально для мобильного приложения при login/refresh/test-login токены дублируются в теле ответа.
    При запросе к защищённым ручкам API мобильному приложению необходимо выставить заголовок Authorization: Bearer *access_token*.
    При refresh мобильное приложение передаёт refresh токен в теле запроса.
  title: Hackathon API
  version: 0.1.0
paths:
...
  /user:
    delete:
      description: Удаляет аккаунт текущего авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Аккаунт успешно удалён
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при удалении аккаунта
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Удалить свой аккаунт
      tags:
      - User
    get:
      description: Получает информацию о текущем авторизованном пользователе. ID берётся
        из JWT токена.
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/User'
              type: object
        "401":
          description: Неверный или отсутствующий токен
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "403":
          description: Неверный формат данных пользователя
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при получении пользователя
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Получить данные текущего пользователя
      tags:
      - User
  /user/{user_id}:
    delete:
      description: Удаляет пользователя по ID. Доступно для пользователей с ролью
        manager и выше.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно удалён
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при удалении пользователя
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Удалить пользователя по ID
      tags:
      - User
    get:
      description: Получает информацию о пользователе по его ID. Доступно без авторизации.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/User'
              type: object
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при получении пользователя
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить пользователя по ID
      tags:
      - User
  /user/block/{user_id}:
    post:
      description: Блокирует пользователя по ID. Доступно для пользователей с ролью
        manager и выше.
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно заблокирован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Неверный параметр пути
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка при блокировке пользователя
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Заблокировать пользователя по ID
      tags:
      - User
  /user/password-forgot:
    post:
      consumes:
      - application/json
      description: Отправляет письмо со ссылкой для восстановления пароля на указанный
        email
      parameters:
      - description: Email пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ссылка для сброса пароля отправлена на email
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Запрос сброса пароля
      tags:
      - User
  /user/password-reset:
    post:
      consumes:
      - application/json
      description: Сбрасывает пароль пользователя по токену. После успешной смены
        токен становится недействительным.
      parameters:
      - description: Данные для сброса пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно изменён
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      security:
      - AccessToken: []
      - RefreshToken: []
      summary: Сброс пароля
      tags:
      - User
swagger: "2.0"